+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - optimized task state determination when 'proceed (y/n)' is present
 - added support for showing untracked files in Updated Files and add to git action
+- added CodeBlock and ExpandableMessageBlock components to extension UI components API
 
 ## [0.78.0]